backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["DavidTrindade98 <dtrindade1998@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.14"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
pydantic = "^1.10.7"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
mysqlclient = "^2.1.1"
mysql = "^0.0.3"
pyjwt = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
